+++ ChangeLog.md
@@ -8,6 +8,7 @@
 
     - Add missing check imports
     - Update emails ans urls translations to be simpler and clearer
+    - Make sure 'modifiedOnUpdate' parameter is fully replaced with 'withCloseButtonWhileNotModified'
 
 ### 1.7.1